fix: changement de l'organisation des dossiers et corrections des espacaes de noms
