+++ Extension/FroozenClockImplementation.php
@@ -1,21 +1,9 @@
 <?php
 
-namespace Koyok\democratia\Extension\Clock;
+namespace Koyok\democratia\Extension;
 
-use DateInterval;
 use DateTimeImmutable;
 use Psr\Clock\ClockInterface;
-
-final class ClockImplementation implements ClockInterface
-{
-    public function now(): DateTimeImmutable
-    {
-        $marge = 60;
-
-        return new DateTimeImmutable('now')->add(new DateInterval('PT2S'));
-
-    }
-}
 
 final class FroozenClockImplementation implements ClockInterface
 {
