+++ src/DeleteMethode.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Koyok\democratia\src;
+
+/**
+ * Méthodes DELETE : Procédures de suppression
+ */
+enum DeleteMethode: string implements Methode
+{
+    case SupprimerInternaute = 'CALL supprimer_internaute(?)';           // 1 param
+}
